+++ Samples~/EasySave Integration/EasySave Integration.cs
@@ -8,6 +8,7 @@ namespace OcDialogue.Samples
 {
     public static class EasySaveIntegration
     {
+        static ES3Settings ES3Settings;
         const string saveDataKey_PresetNames = "Preset";
         const string saveDataKey_DataRowContainer = "DataRowContainer";
         [RuntimeInitializeOnLoadMethod]
@@ -18,6 +19,8 @@ namespace OcDialogue.Samples
                 DBManager.OnRuntimeInitilized += Init;
                 return;
             }
+            ES3.CacheFile();
+            ES3Settings = new ES3Settings(ES3.Location.Cache);
             Load();
             InitSaveCallbacks();
         }
@@ -25,16 +28,16 @@ namespace OcDialogue.Samples
         public static void LoadGameProcessDB(string presetName = "")
         {
             Debug.Log($"Load GameProcessDB | key : {presetName + GameProcessDB.Instance.Address}");
-            if (!ES3.KeyExists(presetName + GameProcessDB.Instance.Address)) return;
+            if (!ES3.KeyExists(presetName + GameProcessDB.Instance.Address, ES3Settings)) return;
             Debug.Log("Overwrite");
-            var saveData = ES3.Load<GameProcessSaveData>(presetName + GameProcessDB.Instance.Address);
+            var saveData = ES3.Load<GameProcessSaveData>(presetName + GameProcessDB.Instance.Address, ES3Settings);
             GameProcessDB.Instance.Overwrite(saveData);
         }
 
         public static void LoadInventory(string presetName = "")
         {
-            if (!ES3.KeyExists(presetName + ItemDatabase.Instance.Address)) return;
-            var saveData = ES3.Load<List<ItemSaveData>>(presetName + ItemDatabase.Instance.Address);
+            if (!ES3.KeyExists(presetName + ItemDatabase.Instance.Address, ES3Settings)) return;
+            var saveData = ES3.Load<List<ItemSaveData>>(presetName + ItemDatabase.Instance.Address, ES3Settings);
             if (Inventory.PlayerInventory == null)
             {
                 Inventory.OnPlayerInventoryChanged += inventory => inventory.Overwrite(saveData);
@@ -47,22 +50,22 @@ namespace OcDialogue.Samples
 
         public static void LoadQuestDB(string presetName = "")
         {
-            if (!ES3.KeyExists(presetName + QuestDB.Instance.Address)) return;
-            var saveData = ES3.Load<List<CommonSaveData>>(presetName + QuestDB.Instance.Address);
+            if (!ES3.KeyExists(presetName + QuestDB.Instance.Address, ES3Settings)) return;
+            var saveData =  ES3.Load<List<CommonSaveData>>(presetName + QuestDB.Instance.Address, ES3Settings);
             QuestDB.Instance.Overwrite(saveData);
         }
 
         public static void LoadNPCDB(string presetName = "")
         {
-            if (!ES3.KeyExists(presetName + NPCDB.Instance.Address)) return;
-            var saveData = ES3.Load<List<CommonSaveData>>(presetName + NPCDB.Instance.Address);
+            if (!ES3.KeyExists(presetName + NPCDB.Instance.Address, ES3Settings)) return;
+            var saveData = ES3.Load<List<CommonSaveData>>(presetName + NPCDB.Instance.Address, ES3Settings);
             NPCDB.Instance.Overwrite(saveData);
         }
 
         public static void LoadEnemyDB(string presetName = "")
         {
-            if (!ES3.KeyExists(presetName + EnemyDB.Instance.Address)) return;
-            var saveData = ES3.Load<List<CommonSaveData>>(presetName + EnemyDB.Instance.Address);
+            if (!ES3.KeyExists(presetName + EnemyDB.Instance.Address, ES3Settings)) return;
+            var saveData = ES3.Load<List<CommonSaveData>>(presetName + EnemyDB.Instance.Address, ES3Settings);
             EnemyDB.Instance.Overwrite(saveData);
         }
         static void Load(string presetName = "")
@@ -80,47 +83,18 @@ namespace OcDialogue.Samples
 
             if (Inventory.PlayerInventory != null)
             {
-                Inventory.PlayerInventory.OnInventoryChanged += () => SaveInventory();
+                Inventory.PlayerInventory.OnInventoryChanged += (item, changeType) => SaveInventory();
             }
             else
             {
-                Inventory.OnPlayerInventoryChanged += inventory => inventory.OnInventoryChanged += () => SaveInventory();
+                Inventory.OnPlayerInventoryChanged += inventory => inventory.OnInventoryChanged += (item, changeType) => SaveInventory();
             }
             
             QuestDB.Instance.OnRuntimeValueChanged += () => SaveQuestDB();
             NPCDB.Instance.OnRuntimeValueChanged += () => SaveNPCDB();
             EnemyDB.Instance.OnRuntimeValueChanged += () => SaveEnemyDB();
         }
-        public static void SaveGameProcessDB(string presetName = "")
-        {
-            ES3.Save(presetName + GameProcessDB.Instance.Address, GameProcessDB.Instance.DataRowContainer.GetSaveData());
-            SavePresetName(DBType.GameProcess, presetName);
-        }
-
-        public static void SaveInventory(string presetName = "")
-        {
-            ES3.Save(presetName + ItemDatabase.Instance.Address, Inventory.PlayerInventory.GetSaveData());
-            SavePresetName(DBType.Item, presetName);
-        }
         
-        public static void SaveQuestDB(string presetName = "")
-        {
-            ES3.Save(presetName + QuestDB.Instance.Address, QuestDB.Instance.GetSaveData());
-            SavePresetName(DBType.Quest, presetName);
-        }
-
-        public static void SaveNPCDB(string presetName = "")
-        {
-            ES3.Save(presetName + NPCDB.Instance.Address, NPCDB.Instance.GetSaveData());
-            SavePresetName(DBType.NPC, presetName);
-        }
-
-        public static void SaveEnemyDB(string presetName = "")
-        {
-            ES3.Save(presetName + EnemyDB.Instance.Address, EnemyDB.Instance.GetSaveData());
-            SavePresetName(DBType.Enemy, presetName);
-        }
-
         /// <summary>
         /// 데이터를 수동으로 저장함.
         /// </summary>
@@ -134,11 +108,46 @@ namespace OcDialogue.Samples
             SaveEnemyDB(presetName);
 
             SavePresetName(presetName);
+            ES3.StoreCachedFile();
+        }
+        public static void SaveGameProcessDB(string presetName = "", bool saveAsFile = false)
+        {
+            ES3.Save(presetName + GameProcessDB.Instance.Address, GameProcessDB.Instance.DataRowContainer.GetSaveData(), ES3Settings);
+            SavePresetName(DBType.GameProcess, presetName);
+            if(saveAsFile) ES3.StoreCachedFile();
+        }
+
+        public static void SaveInventory(string presetName = "", bool saveAsFile = false)
+        {
+            ES3.Save(presetName + ItemDatabase.Instance.Address, Inventory.PlayerInventory.GetSaveData(), ES3Settings);
+            SavePresetName(DBType.Item, presetName);
+            if(saveAsFile) ES3.StoreCachedFile();
+        }
+        
+        public static void SaveQuestDB(string presetName = "", bool saveAsFile = false)
+        {
+            ES3.Save(presetName + QuestDB.Instance.Address, QuestDB.Instance.GetSaveData(), ES3Settings);
+            SavePresetName(DBType.Quest, presetName);
+            if(saveAsFile) ES3.StoreCachedFile();
+        }
+
+        public static void SaveNPCDB(string presetName = "", bool saveAsFile = false)
+        {
+            ES3.Save(presetName + NPCDB.Instance.Address, NPCDB.Instance.GetSaveData(), ES3Settings);
+            SavePresetName(DBType.NPC, presetName);
+            if(saveAsFile) ES3.StoreCachedFile();
+        }
+
+        public static void SaveEnemyDB(string presetName = "", bool saveAsFile = false)
+        {
+            ES3.Save(presetName + EnemyDB.Instance.Address, EnemyDB.Instance.GetSaveData(), ES3Settings);
+            SavePresetName(DBType.Enemy, presetName);
+            if(saveAsFile) ES3.StoreCachedFile();
         }
 
         public static void LoadFromPreset(string presetName)
         {
-            if (!ES3.KeyExists(saveDataKey_PresetNames))
+            if (!ES3.KeyExists(saveDataKey_PresetNames, ES3Settings))
             {
                 Debug.LogWarning($"저장된 프리셋이 없음");
                 return;
@@ -147,7 +156,7 @@ namespace OcDialogue.Samples
         }
         public static void LoadFromPreset(DBType dbType, string presetName)
         {
-            if (!ES3.KeyExists(dbType + saveDataKey_PresetNames))
+            if (!ES3.KeyExists(dbType + saveDataKey_PresetNames, ES3Settings))
             {
                 Debug.LogWarning($"저장된 프리셋이 없음");
                 return;
@@ -175,24 +184,26 @@ namespace OcDialogue.Samples
 
         public static List<string> GetPresetNames()
         {
-            return ES3.KeyExists(saveDataKey_PresetNames) ? ES3.Load<List<string>>(saveDataKey_PresetNames) : new List<string>();
+            return ES3.KeyExists(saveDataKey_PresetNames, ES3Settings) ? 
+                ES3.Load<List<string>>(saveDataKey_PresetNames, ES3Settings) : new List<string>();
         }
         public static List<string> GetPresetNames(DBType dbType)
         {
-            return ES3.KeyExists(dbType + saveDataKey_PresetNames) ? ES3.Load<List<string>>(dbType + saveDataKey_PresetNames) : new List<string>();
+            return ES3.KeyExists(dbType + saveDataKey_PresetNames, ES3Settings) ?
+                ES3.Load<List<string>>(dbType + saveDataKey_PresetNames, ES3Settings) : new List<string>();
         }
 
         public static void RemovePreset(string presetName)
         {
             var keys = GetPresetNames();
             keys.Remove(presetName);
-            ES3.Save(saveDataKey_PresetNames, keys);
+            ES3.Save(saveDataKey_PresetNames, keys, ES3Settings);
         }
         public static void RemovePreset(DBType dbType, string presetName)
         {
             var keys = GetPresetNames(dbType);
             keys.Remove(presetName);
-            ES3.Save(dbType + saveDataKey_PresetNames, keys);
+            ES3.Save(dbType + saveDataKey_PresetNames, keys, ES3Settings);
         }
 
         static void SavePresetName(string presetName)
@@ -202,7 +213,7 @@ namespace OcDialogue.Samples
             if(!existPresets.Contains(presetName))
             {
                 existPresets.Add(presetName);
-                ES3.Save(saveDataKey_PresetNames, existPresets);
+                ES3.Save(saveDataKey_PresetNames, existPresets, ES3Settings);
             }
         }
 
@@ -214,7 +225,7 @@ namespace OcDialogue.Samples
             if (!existPreset.Contains(presetName))
             {
                 existPreset.Add( presetName);
-                ES3.Save(dbType + saveDataKey_PresetNames, existPreset);
+                ES3.Save(dbType + saveDataKey_PresetNames, existPreset, ES3Settings);
                 Debug.Log($"saveKey | {dbType + saveDataKey_PresetNames} = {existPreset[0]}");
             }
         }